refactor: 공통 컴포넌트용 폴더 추가정리
